jimt/CT-822 - FileUploader -> useFileUploader
