+++ apps/src/codebridge/FileBrowser/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useFileUploader';
 export * from './useFileUploadErrorCallback';
 export * from './useHandleFileUpload';
 export * from './usePrompts';
